use nuxtjs/composition-api instead
