+++ client/components/NavText.vue
@@ -25,7 +25,7 @@
 </template>
 
 <script lang="ts">
-import { ref } from '@vue/composition-api'
+import useFixedNav from '~/hooks/useFixedNav'
 // const ModeChange = () => import('~/components/ModeChange.vue')
 
 export default {
@@ -33,22 +33,7 @@ export default {
     // ModeChange
   },
   setup() {
-    const navRef = ref<HTMLDivElement>()
-    if (process.client) {
-      window.onscroll = function updateNav() {
-        if (isPageOffset()) {
-          navRef.value?.classList.add('nav-active')
-        } else {
-          navRef.value?.classList.remove('nav-active')
-        }
-      }
-    }
-    function isPageOffset(): boolean {
-      const offset = navRef.value.offsetTop
-
-      return window.pageYOffset > offset
-    }
-    return { navRef }
+    return { ...useFixedNav() }
   }
 }
 </script>
