+++ client/hooks/useCalendar.ts
@@ -1,4 +1,4 @@
-import { computed } from '@vue/composition-api'
+import { computed } from '@nuxtjs/composition-api'
 import dayjs from 'dayjs'
 
 import { PostItem } from '~/types/blog'
