--- client/plugins/composition-api.ts
@@ -1,4 +1,0 @@
-import Vue from 'vue'
-import CompositionApi from '@vue/composition-api'
-
-Vue.use(CompositionApi)
