+++ package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "@nekohack/footer-ui": "1.0.0",
     "@nekohack/normalize.css": "2.3.7",
+    "@nuxtjs/composition-api": "0.29.0",
     "@nuxtjs/dotenv": "1.4.1",
     "@nuxtjs/google-adsense": "1.4.0",
     "@nuxtjs/google-analytics": "2.4.0",
@@ -33,7 +34,6 @@
     "@nuxtjs/pwa": "3.3.5",
     "@nuxtjs/sentry": "5.1.3",
     "@nuxtjs/sitemap": "2.4.0",
-    "@vue/composition-api": "0.6.7",
     "contentful": "8.5.8",
     "dayjs": "1.10.7",
     "markdown-it": "12.2.0",
